dziala z fordem paliwo wolty i temperatura

+++ arduino/CAR_PC/ELM327.h
@@ -21,6 +21,6 @@ class ELM327
     void connectingToELM327BT(String MAC_ELM327);
     int engineCoolantTemperature();
     double fuelTankLevel();
-    double getVoltage();
+    float getVoltage();
 };
 #endif
